django_api/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi-cli==0.0.5
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mpld3==0.5.10
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.52.2
packaging==24.1
pandas==2.2.3
patsy==1.0.1
pgvector==0.3.6
pillow==11.0.0
pmdarima==2.0.4
protobuf==5.28.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.0
statsmodels==0.14.4
sympy==1.13.1
tbb==2021.11.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
# torch==2.3.1
# torchaudio==2.3.1
# torchvision==0.18.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.1.3